Revert "Covexalgos v2"
