--- .streamlit/config.toml
@@ -1,2 +1,0 @@
-[client]
-showSidebarNavigation = false
