+++ config.properties
@@ -1,21 +1,10 @@
-# 1.Related stocks
-SYMBOLS = MSFT,PLTR,TSLA,NVDA,AMZN,NFLX,BA,GS,SPY,QQQ,IWM,SMH,RSP,AAPL,ABBV,CDNS,INTC,GSPC,v,AIR,PINE,AMX,SPDR,,META,BRK.B,CRWD
+# 0.Related stocks
+SYMBOLS = MSFT,PLTR,TSLA,NVDA,AMZN,NFLX,BA,GS,SPY,QQQ,IWM,SMH,RSP,AAPL,ABBV,CDNS,INTC,GSPC
 
-# 4. Trading Period & Interval
-PERIOD = 5d
-INTERVAL = 5m
-
-# 8. Related metric
+# 2. Related metric
 METRICS = Operating Margin,Dividend Yield,Dividend Cover,Debt/EBITDA,P/E ratio,PEG ratio
 
-# 11. Stop_loss_price / Take_profit_price parameters
+# 6. Stop_loss_price / Take_profit_price parameters
 STOP_LOSS = 1.75
 TAKE_PROFIT = 3
-
-# 15. Trading Strategy List
-MOVING_AVERAGE_BASED = 5/8 SMA,5/8 EMA,5/8 EMA 1-2 candle price
-TREND_BASED = 4-3-1 candle price reversal
-
-MOVING_AVERAGE_WINDOW_SHORT = 5
-MOVING_AVERAGE_WINDOW_LONG = 8
