+++ user_selected_options.csv
@@ -4,6 +4,3 @@ ABBV
 CDNS
 INTC
 NVDA
-TSLA
-AIR
-CRWD
